deleted initializeLogButton.js and JournalEntriesMail.js as they were legacy items no longer required since moving to a JSON file, rather than an email for saving notes. Updated basic styling of entry list elements; refactored css styling to seperate files for each component rather than having all styling in main; updated some comments to be more accurate.
