+++ scripts/JournalEntryForm.js
@@ -1,5 +1,6 @@
 import { saveEntry } from "./journalDataProvider.js";
 
+// Purpose: To render a journal entry form to the DOM.
 const eventHub = document.querySelector(".container");
 const formContainerContentTarget = document.querySelector(".formContainer");
