+++ scripts/JournalEntryList.js
