--- scripts/initializeLogButton.js
@@ -1,6 +1,0 @@
-// import { journalEntries } from "./journalEntriesMail.js";
-
-// export const initializeLogButtonEvent = () => {
-//     const button = document.querySelector("#journalButton");
-//     button.addEventListener("click",journalEntries);
-// };
