+++ scripts/main.js
@@ -1,5 +1,4 @@
-import { entryListComponent } from "./journalEntryList.js";
-// import { initializeLogButtonEvent } from "./initializeLogButton.js";
+import { entryListComponent } from "./JournalEntryList.js";
 import { getEntries } from "./journalDataProvider.js";
 import { JournalEntryForm } from "./JournalEntryForm.js";
 import { LogJournalEntryButton } from "./LogJournalEntryButton.js";
@@ -8,4 +7,3 @@ import "./EditJournalEntryDialog.js"
 JournalEntryForm();
 LogJournalEntryButton();
 getEntries().then(entryListComponent)
-// initializeLogButtonEvent();
