+++ styles/main.css
@@ -1,4 +1,8 @@
 @import './elements/editJournalEntryDialog.css';
+@import './elements/journalEntryButton.css';
+@import './elements/journalEntryForm.css';
+@import './elements/journalEntryList.css';
+
 body {
     background-color: burlywood;
 }
@@ -15,11 +19,6 @@ body {
     top: 110px;   
 }
 
-#concepts {
-    height: 18px;
-    width: 335px;
-}
-
 .container {
     background-image: url("https://www.tagesspiegel.de/images/1496907/24145614/2-format6001.jpg?inIsFirst=true");
     background-repeat: no-repeat;
@@ -29,57 +28,11 @@ body {
     width: 750px;
 }
 
-#date {
-    margin-left: 49px;
-}
-
-#dateLabel {
-    margin-left: 108px;
-}
-
 .flex {
     display: flex;
     flex-wrap: nowrap;
 }
 
-.formContainer {
-    display: inline-block;
-    margin-top: 260px;
-}
-
-
-#journal {
-    width: 745px;
-}
-
-#journalButton {
-    background-color: rgb(212, 166, 79);
-    display: block;
-    font-family: 'Eagle Lake', cursive;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 10px;
-    padding: 1em;
-}
-
-.journalEntry {
-    background-color: whitesmoke;
-    margin-top: 10px;
-}
-
-.journalEntry p {
-    white-space: pre-line;
-}
-
 .labels {
     font-family: 'Share Tech Mono', monospace;
-}
-
-#mood {
-    height: 24px;
-    margin-left: 48px;
-}
-
-#moodLabel {
-    margin-left: 250px;
 }
